[Refactor] trim port, adaptor 패키지 구조 변경 및 rename (#295)
* refactor: url mapping 수정

@RequestMapping 제거

* feat: GetTrimsPort 추가

* feat: GetTrimsAdaptor 추가

* refactor: Port 변경

TrimPort에서 GetTrimsPort로 변겨

* chore: Adaptor 네이밍 변경

* chore: checkStyle 적용
